Added Animation Length slider
Added a slider in Appearance to adjust the length of the open animation for the launcher. By default the animation should be the same as it was previously (the first part of the animation is about 10ms shorter so that the ratio is simpler - 2/3 instead of 25/36)
Currently only has language support for English and the length is adjustable from 100ms to 2000ms in steps of 10.