rework how we fetch plugins from manifest URLs
Introduces the concept of a store of community plugins, which is currently
limited to the official PluginsManifest repository.

Each store can support more than one manifest file URLs. When fetching, all
URLs are used until one of them succeeds. This fixes issues with geo-blocking
such as #2195

Plugin stores can be expanded in the future to be user-configurable, see #2178