Refactor ActionKeywords UI and improve keyword handling
Updated `ActionKeywords.xaml` to use a new `Grid` layout, enhancing the structure and flexibility of the UI. Removed a couple of layers of nested grids and stackpanels. Changed tbOldActionKeyword from a TextBlock to a TextBox so user can copy the text.
Modified `ActionKeywords.xaml.cs` so the old keywords are already filled in, in the TextBox.